Fix a bug where the server crashed when someone reconnected.

+++ game/server/ingame-server/src/main/java/amara/applications/ingame/server/appstates/ServerEntitySystemAppState.java
@@ -134,7 +134,7 @@ public class ServerEntitySystemAppState extends EntitySystemHeadlessAppState<Ing
     }
 
     public void removeInitializedClient(int clientId) {
-        initializedClientIds.remove(clientId);
+        initializedClientIds.remove((Integer) clientId);
         Game game = mainApplication.getGame();
         GamePlayer player = game.getPlayerByClientId(clientId);
         // TODO: If MMO Map
